requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.8
langchain-google-genai>=0.0.6
langchain-text-splitters>=0.0.1
chainlit>=1.0.200
flask>=2.3.3
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
google-generativeai>=0.3.2 